pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Costard86 <Costard86@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^21.2.0"
flask = "^2.0.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^4.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
